+++ tests/nutsml/test_viewer.py
@@ -7,7 +7,7 @@ import numpy as np
 
 from nutsflow import Consume
 from nutsflow.common import Redirect
-from nutsml import PrintColType
+from nutsml import PrintColType, PrintType
 
 expected_col_info1 = """
 item 0: <tuple>
@@ -36,6 +36,12 @@ item 1: <tuple>
   1: <int> 4
 """
 
+expected_type_info = """
+(<ndarray> 3x4:uint8, <ndarray> 1x2:float32)
+<float> 1.1
+[[<ndarray> 3x4:uint8], <int> 2]
+"""
+
 
 def test_PrintColType():
     with Redirect() as out1:
@@ -52,3 +58,12 @@ def test_PrintColType():
         data = [(1, 2), (3, 4)]
         data >> PrintColType((0, 1)) >> Consume()
     assert out3.getvalue() == expected_col_info3[1:]
+
+
+def test_PrintType():
+    a = np.zeros((3, 4), dtype='uint8')
+    b = np.zeros((1, 2), dtype='float32')
+    with Redirect() as out1:
+        data = [(a, b), 1.1, [[a], 2]]
+        data >> PrintType() >> Consume()
+    assert out1.getvalue() == expected_type_info[1:]